Add support for configuring S3 bucket logging.
All logs will go to a single bucket per account. The key prefix for
logs in this bucket will be determined by
<appName>-<envName>-<serviceName>
