+++ lib/common/s3-static-site-logging-bucket.yml
