+++ lib/services/s3/index.js
@@ -20,6 +20,7 @@ const accountConfig = require('../../common/account-config')().getAccountConfig(
 const cloudFormationCalls = require('../../aws/cloudformation-calls');
 const deployPhaseCommon = require('../../common/deploy-phase-common');
 const deletePhasesCommon = require('../../common/delete-phases-common');
+const s3DeployersCommon = require('../../common/s3-deployers-common');
 const preDeployPhaseCommon = require('../../common/pre-deploy-phase-common');
 const bindPhaseCommon = require('../../common/bind-phase-common');
 const handlebarsUtils = require('../../common/handlebars-utils');
@@ -69,7 +70,7 @@ function getDeployContext(serviceContext, cfStack) {
 }
 
 
-function getCompiledS3Template(stackName, ownServiceContext) {
+function getCompiledS3Template(stackName, ownServiceContext, loggingBucketName) {
     let serviceParams = ownServiceContext.params;
 
     let bucketName = serviceParams.bucket_name || stackName;
@@ -83,6 +84,11 @@ function getCompiledS3Template(stackName, ownServiceContext) {
         versioningStatus: versioningStatus,
         tags: deployPhaseCommon.getTags(ownServiceContext)
     };
+
+    if(serviceParams.logging && serviceParams.logging === 'enabled') {
+        handlebarsParams.loggingBucketName = loggingBucketName;
+        handlebarsParams.logFilePrefix = s3DeployersCommon.getLogFilePrefix(ownServiceContext);
+    }
 
     return handlebarsUtils.compileTemplate(`${__dirname}/s3-template.yml`, handlebarsParams)
 }
@@ -116,7 +122,10 @@ exports.deploy = function (ownServiceContext, ownPreDeployContext, dependenciesD
     let stackName = deployPhaseCommon.getResourceName(ownServiceContext);
     winston.info(`${SERVICE_NAME} - Deploying S3 bucket ${stackName}`);
 
-    return getCompiledS3Template(stackName, ownServiceContext)
+    return s3DeployersCommon.createLoggingBucketIfNotExists()
+        .then(loggingBucketName => {
+            return getCompiledS3Template(stackName, ownServiceContext, loggingBucketName)
+        })
         .then(compiledTemplate => {
             let stackTags = deployPhaseCommon.getTags(ownServiceContext);
             return deployPhaseCommon.deployCloudFormationStack(stackName, compiledTemplate, [], true, SERVICE_NAME, stackTags);
