+++ lib/services/s3staticsite/index.js
@@ -20,9 +20,9 @@ const DeployContext = require('../../datatypes/deploy-context');
 const cloudFormationCalls = require('../../aws/cloudformation-calls');
 const deployPhaseCommon = require('../../common/deploy-phase-common');
 const deletePhasesCommon = require('../../common/delete-phases-common');
+const s3DeployersCommon = require('../../common/s3-deployers-common');
 const preDeployPhaseCommon = require('../../common/pre-deploy-phase-common');
 const bindPhaseCommon = require('../../common/bind-phase-common');
-const accountConfig = require('../../common/account-config')().getAccountConfig();
 const handlebarsUtils = require('../../common/handlebars-utils');
 
 const SERVICE_NAME = "S3 Static Site";
@@ -36,7 +36,7 @@ function getCompiledS3Template(ownServiceContext, stackName, loggingBucketName) 
 
     let bucketName = serviceParams.bucket_name || stackName;
     let versioningStatus = VERSIONING_PARAM_MAPPING[serviceParams.versioning] || 'Suspended';
-    let logFilePrefix = `${ownServiceContext.appName}/${ownServiceContext.environmentName}/${ownServiceContext.serviceName}/`;
+    let logFilePrefix = s3DeployersCommon.getLogFilePrefix(ownServiceContext);
     let indexDocument = serviceParams.index_document || 'index.html';
     let errorDocument = serviceParams.error_document || 'error.html';
 
@@ -51,22 +51,6 @@ function getCompiledS3Template(ownServiceContext, stackName, loggingBucketName) 
     };
 
     return handlebarsUtils.compileTemplate(`${__dirname}/s3-static-site-template.yml`, handlebarsParams)
-}
-
-function createLoggingBucketIfNotExists() {
-    let stackName = "HandelStaticSiteLoggingBucket";
-    let bucketName = `handel-static-site-logging-${accountConfig.region}-${accountConfig.account_id}`;
-    let handlebarsParams = {
-        bucketName
-    }
-
-    return handlebarsUtils.compileTemplate(`${__dirname}/s3-static-site-logging-bucket.yml`, handlebarsParams)
-        .then(compiledTemplate => {
-            return deployPhaseCommon.deployCloudFormationStack(stackName, compiledTemplate, [], false, "Logging Bucket for S3 Static Sites");
-        })
-        .then(deployedStack => {
-            return cloudFormationCalls.getOutput("BucketName", deployedStack);
-        });
 }
 
 /**
@@ -98,7 +82,7 @@ exports.deploy = function (ownServiceContext, ownPreDeployContext, dependenciesD
     let stackName = deployPhaseCommon.getResourceName(ownServiceContext);
     winston.info(`${SERVICE_NAME} - Deploying S3 static website '${stackName}'`);
 
-    return createLoggingBucketIfNotExists()
+    return s3DeployersCommon.createLoggingBucketIfNotExists()
         .then(loggingBucketName => {
             return getCompiledS3Template(ownServiceContext, stackName, loggingBucketName)
         })
